[refactor]: custom 에러로 throw
